backend/test_global_admin.py
#!/usr/bin/env python3
"""
Test script for global admin creation functionality
"""

import sys
import os
import requests
import json

# This script should be run with the server running
BASE_URL = "http://localhost:8000"

def test_global_admin_creation():
    print("Testing global admin creation functionality...")
    
    # Note: This test requires a global admin account to already exist
    # In a real scenario, you'd need to:
    # 1. Have at least one global admin account
    # 2. Login as that admin to get a token
    # 3. Use that token to create another global admin
    
    print("""
    To test the global admin creation feature:
    
    1. Start the backend server: python server.py
    2. Start the frontend: npm start
    3. Login as a global admin user (role = 1)
    4. Navigate to Team Management
    5. Click on the "Global Admins" tab
    6. Use the "Create Global Admin" button
    
    The new features include:
    - Secure endpoint: /api/admin/create-global-admin (requires global admin token)
    - New UI tab in Team Management for global admin management
    - Form to create new global administrators
    - List of current global administrators
    - Proper validation and permission checking
    
    Backend changes:
    - Added secure endpoint that validates requester is global admin
    - Updated register endpoint to only create regular users
    - Added proper JWT token validation for admin functions
    
    Frontend changes:
    - Added "Global Admins" tab (only visible to global admins)
    - Added form to create global admins with password generation
    - Added list view of current global administrators
    - Added appropriate styling with red theme for admin functions
    """)

if __name__ == "__main__":
    test_global_admin_creation()
