backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
openpyxl
aiofiles
chromadb
openai
qdrant-client
python-docx
PyMuPDF
Pillow
tiktoken
tenacity
PyJWT
aiosmtplib
open-clip-torch
torch
torchvision
flask